Beginning of Effect Implementation
I've added the bones for the ability implementation. I've decided to implement it in such a way that each card effect will require a specific function. The abilities will interact with each other through the use the of an event queue and stack while using a new class called CardEvents these will provide the backbone of all actions taken in the game.