SQL Database Implementation
A SQL Database has been added to the project to replace the old text file. As such the CardDatabase Static class is currently being rewritten to utilize the DB. Alot of things are now broken because of this change. No portion of deck building is still operational so that will be rebuild after the abilities and traits are taken care of. Currently the base classes constructors are being reconfigured to mesh nicely with the DB. 1/20/19 10hrs