Setting up the Event System
I have more or less overhauled the Location class to better describe the actions in the game. I have created a Turn class that contains enum turn phases so that the EventManager can easily control turn order and dimension twisting. The backbone of the event system is being laid, the GameEventArgs class is being fleshed out so that it is robust enough to accommidate the entire game but not so cumbersome that the code's just shit. Spent 6hr