Base Classes are Done
I have finished the base classes. And more or less completed the SQL database implementation. There's the bug that if there are NULLs in the middle instead of the just the end. It gets wonky. I have started rebuilding the deck builder, the plan is to include functionality that saves the player deck to the Database. Infact thats where all player data will be saved.