Base Classes Tested
I have heavily modified the GameEventsManager script that it "runs through a game". It just reduces a Majesty's HP whenever the space bar is pressed, but this allows me to make sure turns work properly. The plan for Thursday is to implement the Turn phase functions that automatically do the things required for the specific turn phase. The weekend goal is to finish the deck building, implement deck saving in the database. Next week and beyond is to work out a demo interface, fix the display cards script and actually start working on abilities. Then once Abilities for Mytics are done, the goal will be to finialize everything and the real fun starts with UI design and things. Time spent  5hrs 45min.