Onward to fixing the DisplayFunction and Implementing the Prototype UI
As much as possible for the Turn phases has been completed, while the displayFunction remains broken and there being no gameplay UI. I have implemented Territory Challenge, fixed some things in the Trait constructor and general turn things. There is a non-bug with the dynamic card creation using the Activator the issue is that certains strings end up in the wrong place for Philosophers their ability string from the DB ends up as their trait which gets split by the commas. The Actions such as Attack, Block and activation need to be made into a CardAction Class. This will take place after the displayfunction is fixed and the general gameplay UI is built. Once the CardAction class is completed the abilities will be worked on. Once the  abilities are implemented for the Mytics, networking will be added, a build will be made for the team and work on AI will begin. Paired with other Kazoku implementations. ALSO there's an error with the Drawing it appears based on debug that players are drawing from each other's hands. So fix that Time spent 5hrs