Fixed Deck Building
After the over hual to the structure of the card's base classes. The previous deck building code broke. It is now fixed and more efficient. No longer do those 2 scripts exist that purely store Card data. The deck builder has also been updated to save the deck you make in the database. Before the displayFunction is fixed and prototype UI is built. The only things left is to visually populate the saved deck when the user starts editing their deck.