Making the Abilties
A function that reads from a selected table has been added to HelperFunctions (its also supposed to write and issue commands but that hasn't been implemented yet, because I don't need it right now). I have decided to go the automation routine for the abilities. What needs to be done is that all card abilities will need to be broken down into their base functions. A Function should be made for those base functions and the ability itself will be put together on the fly. On card creatation which only happens once. Also the game will have Mixer integration.