Database Updates
Updated the AbilityEncodes table as per the eencoded.txt file in the resources folder of the project.  The format is slightly different in the database than the file. Left out the sources because I feel they present an unneeded extra challenge at this point. Next is to create the conditions and actions