TimeKeeping In Game
For a while Ive had the idea to run the game like a state machine or a piece of hardware with a clock signal. Ive begun implementing a GameTimeManaget class which keeps time and performs the one time setup for each card game players, TC's, setting turn order. Its being tied into the Turn class because each turn phase is one unit of gametime. Currently alot of test stuff that was in the GameManager class is being moved to GameTimeManager I think it fits better. In a test scene there's some UI that needs to have some if its the player's turn logic implemented. The plan is to build up this time keeping, then test some compile code on runtime stuff and get to the abilities going. The goal is to have the abilities written in such a way that its not every difficult for the designers to write them once they learn the format plus the tool I create will have compiling in it.