Continue GameTimeManager and Make Additions to DisplayCards
I made a function in Utilities namespace for displaycards requires less typing. There also needs to be a filter added to choose what type of cards are selected. A Criteria object that searches the source for things that match. There needs to be a AI namespace where function calls are made to its exposed bit for the AI to choose a card for territory challenge and such. Thats more of a code as you go thing. The Player class needs to be documented and will turn into a place where functions exist that encompass all a player can do. It will just  be handled differently based on whether or not the Player is an AI or Networked (Future addition) But start continue working on the GameTimeManager