Bug Fixes
Mostly I fixed alot of bugs with the cards although there is still the problem of cards being moved into other players locations. Additionally the Renderer is being worked on with the UI and then the cards in hand. Another things is to slow down the AI.
Next Work: Continue with renderer, cards in hand, and UI plus deck bug fixes