Back To Progress
THe last commit definitely broke a bunch of stuff thats all fixed now. The RenderManager is NOT complete it doesnt display the cards in the players hand let alone update them. The code exists it just doesnt work? Not sure thats the next thing to get working. There was also a static EventProcessor class created to process events as they hit the queue. This is where animations will play Im thinking of sending the info for UI changes through event processor because it makes the most sense.