Beginning Refactoring working on the GameEvents
I have begun implementing interfaces, starting with using the Observer pattern and Command pattern with the handling of GameEvents. There will also be changes made to a few other classes with the use of the Strategy Pattern. The Observer pattern is pretty much implemented in the GameEvent and GameEventManager classes. Next is for continue researching Lambda expressions for simplified implementation of the Command pattern for doing the abilities/events.