Reworking Some Things
This push is definitely not gonna compile but I know that. The PlayerInteraction, Location, Turn, and GameManager classes need to be updated for the new changes described in the [developer notes](https://docs.google.com/document/d/1exB_kNYvD8-NJ7PrPmzgC-jZWdG5x3Tvh-NlASAUhUQ/edit?usp=sharing). After these additions the turns will be fixed to link with UI and you should be able to work through a turn with cardPlaying and UI updates and all by the commit on Tuesday.