Made it Nice
The changes were reverted but I made some new changes and fixed some of the bugs that were created. The next thing I want to do is work on the Turns/Gameplay. Which means first fixing the bug where Majesty HP is zero which is weird but the code is cleaner now