It WORKS!
I have finally actually completed the turn phases control human and Ai everything. There are a few bugs such as Awakening isnt just Philosophers thats a problem with the DSC.
Now I can actually work on the Attack functionality.
The overlay is also ugly and hard to read but these are beautify things. WE DID IT!