UI Events are coming along
It is now possible to draw cards and play them. The cards in the hand are properly spaced the ones on the field are not. The UIManager needs to use Observable Collections so that it can use events/callbacks to update the position of cards on the field and hand when the field and hand collects are changed. Also the events are flowing through the transport layer and are being checked for legality by the game manager.

Cardscript.cs and CardControl.cs have been merged
GameEventsManager.cs and GameMaster.cs will be merged at some point